requirements.txt
pytest==5.4.3
pytest-html==2.1.1
allure-pytest==2.8.16
PyYAML==5.3.1
selenium==3.141.0
zmail==0.2.8
